Consolidate models into utils
