+++ src/index.ts
@@ -1,2 +1,2 @@
 export { calculate, derange as calculateSync } from './derange';
-export { DerangementError, Person, Exclusion } from './models';
+export { DerangementError, Exclusion, Person } from './utils';
